Switch from using HeapOperandFrom() to MemOperand and fix bug
The bug was in CodeGeneratorARM64::GenerateRestoreStub(). Previously, the
generated code was moving the stub address into the object register before
zeroing out the 4 bytes of memory at offset 12 from the address in the object
register. As a result, the code was zeroing out bytes in the stub header
instead of the object header. This bug didn't cause any visible errors, because
the 4 bytes of memory at offset 12 in the stub header are currently just
padding. Now the generated code correctly zeroes out the memory in the object
header before moving the stub address into the object register.