Add reclamation table entry usage to interpreter and compiler
This commit changes the swap system so that each stub holds a pointer to
a reclamation table entry instead of directly pointing to its corresponding
object. This commit changes both the interpreter and compiler to take this
added layer of indirection into account and to perform reclamation table entry
locking/unlocking around accesses to a stub's object.

The compiler modifications in this commit introduce a few race conditions that
could appear if the OS is actively harvesting memory from a running app. Some
are caused by the VIXL assembler's lack of support for atomic instructions and
could be easily fixed if that support were added. One race condition is
trickier, though: because only one temp register is guaranteed to be available
when GenerateRestoreStub() is called, that method has to unlock the reclamation
table entry before reading the stub address from the object header and putting
it in the object address.  I'm not sure if there is any way to fix this race
condition given that we don't have a second temp register available.