Change compiler to reserve register for unlocking RTEs
Previously, the compiler saved back-pointers inside of the object
header and reclamation table entry and used those back-pointers to
unlock the RTE when restoring the stub address into the object
register. This strategy meant that compiled code would not unlock the
RTE when the object register was overwritten with the output of an
object access. This commit changes the compiler to reserve register r20
and save the stub address in that register during every object access.
Compiled code then uses the saved stub address to unlock the RTE, so
that it can unlock the RTE even when the object register is overwritten.