Change swap code to use stub pointers in all data structures
This commit changes the swap code to use stub pointers instead of pointers to
the objects themselves when storing information about objects with stubs in the
swapOutSet, writeQueue, and writeSet data structures. The point of this commit
is to make sure that the swap code never holds onto a pointer to an object for
which a stub exists, since according to our design, that object could be
reclaimed by the OS at any time. This commit does not include the reclamation
table entry locking and unlocking required by our design, but it should make it
easier to add thath locking and unlocking.

These changes mean that SwapObjectsOut() and writeToSwapFile() need to check if
the objects with which they are working are actually stubs. On the bright side,
these changes mean that the objectStubMap and all its associated bookkeeping
are no longer necessary. We still need some way of keeping track of stubs and
their associated objects during the semi-space GC, so I added a stubObjectMap
to hold that information, but the fact that we only need that information
during the semi-space GC means that the bookkeeping associated with the new map
is much simpler.