Change swap code to avoid going through memory allocator
This commit changes the swap code to "reclaim" objects in SwapObjectsOut() by
directly calling madvise() on their pages rather than freeing them from the
swappedInSpace, avoiding the need to re-allocate space for the objects when
swapping them back in. Because the reclaimed objects now stay at the same
virtual addresses when they are swapped back in, this commit also adds a
"resident bit" to the reclamation table entry and changes all code that checks
if an object has been reclaimed to check if the resident bit is cleared instead
of checking if the RTE's object address is null.